requirements.txt
# Enhanced EUR/USD Trading Signal Bot
# Python Version: Minimum 3.8, Recommended 3.12, Tested with 3.8-3.13

# Core dependencies
aiohttp>=3.10.0
aiogram>=3.13.0
python-dotenv>=1.0.0
APScheduler>=3.10.4

# Data processing
pandas>=2.2.2

# Technical indicators
ta==0.11.0

# GPT / OpenAI
openai>=1.54.0
httpx>=0.27.0

# Database
aiosqlite>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

